requirements.txt
moderngl==5.11.1
numpy==2.1.1
pygame==2.6.0
PyGLM==2.7.2
pillow==10.4.0
PyOpenGL==3.1.7
imgui_bundle==1.5.2